Include libgirepository libraries (#215)
* include libgirepository libraries

* fix lib path

* add libgirepo

+++ appimage/appimagegen.sh
@@ -64,6 +64,24 @@ for LIBOSM_PATH in $LIBOSM_PATHS; do
 done
 rm -rf squashfs-root
 
+
+# Ensure destination exists
+mkdir -p "$APP_DIR/usr/lib"
+
+# Find libgirepository libraries and copy them to your AppDir
+LIBGIREPO_PATHS=$(ldconfig -p | grep libgirepository | awk '{print $4}' | sort -u)
+
+if [[ -z "$LIBGIREPO_PATHS" ]]; then
+    echo "libgirepository libraries not found on the system, exiting."
+    exit 1
+fi
+
+for libpath in $LIBGIREPO_PATHS; do
+    echo "Copying $libpath to $APP_DIR/usr/lib/"
+    cp "$libpath" "$APP_DIR/usr/lib/"
+done
+
+
 # 5. Modify the AppRun file to add LD_LIBRARY_PATH after the gtk plugin line
 echo "Configuring AppRun file..."
 rm "$APP_DIR/AppRun"
